Make RF driver rand seeding conditional (#8)
Add a heuristic to avoid calling RF driver calls that may not
be present. We try to check if Nanostack is present by looking for
MBED_CONF_NANOSTACK_CONFIGURATION being defined - this should
indicate that it's in the build, and hence there should be a radio
driver.

Not perfect, but should get some build attempts unstuck.